made some implementation changes to the home layout
