+++ src/components/map.js
@@ -6,10 +6,9 @@ import mapboxgl from "mapbox-gl";
 
 mapboxgl.accessToken = "pk.eyJ1IjoidHdhYmkiLCJhIjoiY2tlZnZyMWozMHRqdjJzb3k2YzlxZnloYSJ9.FBL3kyXAQ22kEws-y6XbJQ";
 
-const Map = () => {
+const Map = (props) => {
 
     const mapContainerRef = React.useRef(null);
-    const [locationText, setLocationText] = React.useState("");
 
     const getLocationNames = (lat, long) => {
         var loc = "";
@@ -27,7 +26,7 @@ const Map = () => {
                 var lity = result.locality;
 
                 loc = country +", "+sub + ", "+lity;
-                setLocationText(loc);
+                props.getLocation(loc);
                 return loc;
             }).catch((error) => {
             alert("oops an error occurred: " + error + " .Try reloading your page");
@@ -38,6 +37,10 @@ const Map = () => {
 
     // initialize map when component mounts
     React.useEffect(() => {
+
+        //if(!props.isLoggedIn){
+        //window.location.href = "/";
+        //}
 
         navigator.geolocation.getCurrentPosition((position) => {
             const userCoordinates = [position.coords.longitude, position.coords.latitude];
@@ -52,10 +55,11 @@ const Map = () => {
 
 
             map.on("load", function() {
+                props.popupCallBack();
                 map.loadImage(myIcon,
                     function(error, image) {
                         map.addImage("myIcon-marker", image);
-                });
+                    });
                 map.loadImage(
                     mechIcon,
                     function(error, image) {
@@ -78,6 +82,52 @@ const Map = () => {
                             }
                         });
 
+                        //dummy location data and mechanic locations
+                        var geojson = {
+                            "type": "FeatureCollection",
+                            "features": [
+                                {
+                                    "type": "Feature",
+                                    "properties": {
+                                        "message": "Foo",
+                                        "iconSize": [60, 60]
+                                    },
+                                    "geometry": {
+                                        "type": "Point",
+                                        "coordinates": [position.coords.longitude+0.004, position.coords.latitude+0.006]
+                                    }
+                                },
+                                {
+                                    "type": "Feature",
+                                    "properties": {
+                                        "message": "Bar",
+                                        "iconSize": [50, 50]
+                                    },
+                                    "geometry": {
+                                        "type": "Point",
+                                        "coordinates": [position.coords.longitude+0.009, position.coords.latitude+0.003]
+                                    }
+                                },
+                                {
+                                    "type": "Feature",
+                                    "properties": {
+                                        "message": "Baz",
+                                        "iconSize": [40, 40]
+                                    },
+                                    "geometry": {
+                                        "type": "Point",
+                                        "coordinates": [position.coords.longitude+0.005, position.coords.latitude+0.002]
+                                    }
+                                }
+                            ]
+                        };
+
+
+                        map.addSource("mechPoints", {
+                            "type": "geojson",
+                            "data": geojson
+                        });
+
                         map.addLayer({
                             "id": "symbols",
                             "type": "symbol",
@@ -86,10 +136,39 @@ const Map = () => {
                                 "icon-image": "myIcon-marker"
                             }
                         });
+                        map.addLayer({
+                            "id": "mechSymbols",
+                            "type": "symbol",
+                            "source": "mechPoints",
+                            "layout": {
+                                "icon-image": "custom-marker"
+                            }
+                        });
 
                         map.flyTo({
                             center: userCoordinates,
-                            zoom: 14
+                            zoom: 15
+                        });
+
+                        // Center the map on the coordinates of any clicked symbol from the 'symbols' layer.
+                        map.on("click", "mechSymbols", function(e) {
+                            map.flyTo({
+                                center: e.features[0].geometry.coordinates
+                            });
+                            new mapboxgl.Popup()
+                                .setLngLat(e.features[0].geometry.coordinates)
+                                .setHTML("<h3>title</h3><p>some mechanic</p>")
+                                .addTo(map);
+                            alert("some mechanic");
+                        });
+
+                        map.on("mouseenter", "mechSymbols", function() {
+                            map.getCanvas().style.cursor = "pointer";
+                            //alert("some mechanic");
+                        });
+
+                        map.on("mouseleave", "mechSymbols", function() {
+                            map.getCanvas().style.cursor = "";
                         });
 
                         // Center the map on the coordinates of any clicked symbol from the 'symbols' layer.
@@ -107,7 +186,7 @@ const Map = () => {
                         map.on("mouseenter", "symbols", function() {
                             // Change the cursor style as a UI indicator.
                             map.getCanvas().style.cursor = "pointer";
-                            alert("your position");
+                            //alert("your position");
 
                         });
